迎新纳福 红包入袋

+++ Core/GObject/Player.h
@@ -2105,7 +2105,7 @@ namespace GObject
         void callSnakeEgg();
         void getSnakeEggAward(UInt8 v);
 
-        void getNewYearGiveGiftAward(UInt8 dayOrder, UInt16 result);
+        void getNewYearGiveGiftAward(UInt8 dayOrder, UInt8 result);
     private:
         UInt8 cf_posPut[5];//范围1-5
         UInt32 cf_itemId[5];
